Add masking option in reconstruction plugin (#268)
The actual reconstruction plugin checks if the Acquisition has checkbox "use mask" checked or not to perform a masked/unmasked reconstruction. To modify the parameter, one needs to select Acquisition object and check/uncheck use mask. This commit adds the checkbox option in the plugin interface to avoid going back and forth between usReconstructionPlugin and USAcquisitionObject.