Sequence IO (#419)
Description:
Plugin to import/export US sequences as .igs.mha compatible with (PLUS format). This contribution is meant to help exchanging data between IBIS and 3DSlicer.

For the moment, most parameters in the header are not used or assumed to be fixed by IBIS hence they are hardcoded (AnatomicalOrientation, BinaryData, BinaryDataByteOrderMSB, CenterOfRotation, ElementSpacing, ElementType, Offset, TransformMatrix and UltrasoundImageOrientation).

Two parameters (not used by PLUS) are added to the header:

CalibrationTransform: contains the calibration matrix, if not present calibration is assumed to be identity.
TimestampBaseline: contains the timestamp of the first frame (this value is added to Seq_FrameXXXXXXXXXX_Timestamp to compute the original timestamp in IBIS), if not present TimestampBaseline is assumed to be 0.
Export:
USAcquisitionObjects are exported to igs.mha with the following transforms:

ImageToProbeTransform: contains the calibration matrix as a sequence. This is supposed to be the same along the sequence, and is useful when applying/removing calibration transform with 3DSlicer.
ProbeToReferenceTransform: contains the tracking matrix as a sequence without calibration.
ImageToReferenceTransform: contains the tracking matrix as a sequence with calibration. This is the result of applying ImageToProbeTransform x ProbeToReferenceTransform. It is used for convenience.
Import:
The plugin can load compressed and uncompressed sequences. When importing a .igs.mha sequence, the plugin parses the input file and shows the available transform matrices that can be used as the probe tracking transforms. If the sequence file was generated with IBIS, ProbeToReferenceTransform will be recommended. If CalibrationTransform field is found, the calibration matrix is automatically loaded. It can also be loaded manually as a .xfm file.

TODO:

Export compressed sequences, for now, it only exports in uncompressed format
Export RGB images for color doppler acquisitions
Encode depth parameter of US probe in both igs.mha and in minc files