Restrict Superbuild CMake to single configuration
On Windows the default value for CMAKE_CONFIGURATION_TYPES is
Debug;Release;MinSizeRel;RelWithDebInfo, which causes problems with
WinDeployQt, as outlined in #267. The Ibis Wiki was already requesting
Windows users to set a single configuration, but it wasn't enforced in
the CMakeLists, which led to failed builds when the user didn't comply
with the build instructions. This commits reconciles the wiki
instructions with the CMakeLists. It makes the build process more
straightforward and intuitiuve. Specifically:

* On the first run, the configuration defaults to Release

* The cache is only allowed Debug and Release (single config)

* The CMAKE_BUILD_TYPE variable is automatically set on Windows to
  match the CMAKE_CONFIGURATION_TYPES variable, to enable OpenCV to
  build without requiring the user to set it directly